{{cookiecutter.project_name}}/requirements/base.txt
python-decouple==2.3
raven==5.5.0
scrapy-sentry==0.6.1
Scrapy==1.0.2
scrapyd==1.1.0
scrapyd-client==1.0.1
service-identity==14.0.0
Unipath==1.1
wheel==0.23.0
unicodecsv==0.13.0
psutil==3.2.0
ipdb==0.8.1
ipython==3.2.1
python-scrapyd-api==0.2.0
requests==2.6.0
tox==2.1.1
html2text==2015.6.21
tablib==0.10.0
